+++ package.json
@@ -10,8 +10,7 @@
   },
   "dependencies": {
     "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "tslib": "^2.6.2"
+    "react-dom": "18.2.0"
   },
   "devDependencies": {
     "@babel/core": "^7.14.5",
@@ -46,6 +45,7 @@
     "jsdom": "~22.1.0",
     "nx": "18.2.3",
     "prettier": "^2.8.8",
+    "tslib": "^2.6.2",
     "typescript": "~5.4.5",
     "vite": "~5.0.13",
     "vite-plugin-dts": "~3.8.1",
